Use mock.patch marker to force consistent time stamp
Testing of new block creation requires that the block content remain
exactly the same or we wouldn’t be able to compare the final hash. I
added an @patch right before the test case to force time.time() to
always return the same value when the test is run. I was then able
assertEqual the created hash to a constant string.