Progress and some clean up
We now test the loading of a class from bytes into an object. We store
the binary block in a file in one test, then re-created a CBOject from
the bytes read in from the file.
Deleted the old GPObject related files
Deleted test_json.py.
Added *.dat to ignored files.