Added exception for b58 decode check failure
hashes.py - if check fails, decode logs an error and raises
ResourceWarning with the b58 check address as an argument.
Added test_b58_check_failure test to test_hashes.
wallets.py now does the b58 decode check call inside a try-except. On
the exception, it logs an error and re-raises the exception. This is in
the findAddress method when it’s passed a string as the address. If the
type of the address input arg is not str or bytes, this method raises a
ResourceWarning passing the provide address as an argument.
test_wallets.py has a new test case test_badB58Address test case.